requirements.txt
certifi==2020.11.8
chardet==3.0.4
dataclasses==0.6
dataclasses-json==0.5.2
Deprecated==1.2.10
docker==4.3.1
idna==2.10
marshmallow==3.9.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
Pillow==8.0.1
pygltflib==1.13.6
requests==2.25.0
six==1.15.0
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
websocket-client==0.57.0
wrapt==1.12.1
